Add delegating `AsyncCloseable` and `ListenableAsyncCloseable` classes
Motivation:

It's a frequent case that users need to wrap `AsyncCloseable` objects.
Having delegating implementations will be handy to avoid boilerplate.

Modifications:

- Add `DelegatingAsyncCloseable` class.
- Add `DelegatingListenableAsyncCloseable` class.
- Use these two new classes as a parent class for all existing
delegating classes that also implement `AsyncCloseable` interface.

Result:

1. Users will have handy utilities to delegate async closeable methods.
2. Less code to maintain our existing delegating classes.
