+++ servicetalk-concurrent-api/src/main/java/io/servicetalk/concurrent/api/DelegatingExecutor.java
@@ -23,12 +23,10 @@ import java.util.concurrent.RejectedExecutionException;
 import java.util.concurrent.TimeUnit;
 import java.util.function.Supplier;
 
-import static java.util.Objects.requireNonNull;
-
 /**
  * An {@link Executor} that simply delegates all calls to another {@link Executor}.
  */
-public abstract class DelegatingExecutor implements Executor {
+public abstract class DelegatingExecutor extends DelegatingListenableAsyncCloseable implements Executor {
 
     private final Executor delegate;
 
@@ -38,12 +36,8 @@ public abstract class DelegatingExecutor implements Executor {
      * @param delegate {@link Executor} to delegate all calls to.
      */
     protected DelegatingExecutor(final Executor delegate) {
-        this.delegate = requireNonNull(delegate);
-    }
-
-    @Override
-    public String toString() {
-        return this.getClass().getSimpleName() + "{delegate=" + delegate() + "}";
+        super(delegate);
+        this.delegate = delegate;
     }
 
     /**
@@ -51,6 +45,7 @@ public abstract class DelegatingExecutor implements Executor {
      *
      * @return The delegate {@link Executor} used.
      */
+    @Override
     protected Executor delegate() {
         return delegate;
     }
@@ -104,25 +99,5 @@ public abstract class DelegatingExecutor implements Executor {
     @Override
     public long currentTime(TimeUnit unit) {
         return delegate.currentTime(unit);
-    }
-
-    @Override
-    public Completable onClose() {
-        return delegate.onClose();
-    }
-
-    @Override
-    public Completable onClosing() {
-        return delegate.onClosing();
-    }
-
-    @Override
-    public Completable closeAsync() {
-        return delegate.closeAsync();
-    }
-
-    @Override
-    public Completable closeAsyncGracefully() {
-        return delegate.closeAsyncGracefully();
     }
 }
